==> 2. kosdaq_결측치.ipynb <==
import pandas as pd
import numpy as np
# pd.set_option('display.max_rows',None)

# 데이터로드
kosdaq = pd.read_csv('./datasets/1.kosdaq_preprocessing완료.csv', index_col=0)
kosdaq

# 결과창 늘려주는 코드 시작
pd.set_option('display.max_rows',None)
# 결측치 확인
kosdaq.isna().sum()

# 결과창 늘려주는 코드 종료
pd.reset_option('display.max_rows', None)

# 상장폐지기업 확인
kosdaq['label'].value_counts()

# 성장성 결측치 확인 - 526개
kosdaq[kosdaq['총자본증가율'].isna()]

kosdaq[kosdaq['총자본증가율'].isna()]['label'].value_counts()

# 성장성 결측치 제거 - 528개 제거
kosdaq_na = kosdaq[~kosdaq['총자본증가율'].isna()]
kosdaq_na.shape

# 수익성 결측치 확인 - 1개
kosdaq_na[kosdaq_na['매출액총이익률'].isna()].iloc[:,13:]

# 수익성 결측치 제거 - 1개 제거
kosdaq_na = kosdaq_na[~kosdaq_na['매출액총이익률'].isna()]
kosdaq_na.shape

# 안정성 결측치 확인 - 1개
kosdaq_na[kosdaq_na['유동자산구성비율'].isna()].iloc[:,28:]

# 안정성 결측치 제거 - 1개 제거
kosdaq_na = kosdaq_na[~kosdaq_na['유동자산구성비율'].isna()]
kosdaq_na.shape

# 활동성 결측치 확인 - 1개
kosdaq_na[kosdaq_na['총자본회전률'].isna()].iloc[:,49:]

# 활동성 결측치 제거 - 1개 제거
kosdaq_na = kosdaq_na[~kosdaq_na['총자본회전률'].isna()]
kosdaq_na.shape

# 결과창 늘려주는 코드 시작
pd.set_option('display.max_rows',None)
# 결측치 확인
kosdaq_na.isna().sum()

# 결과창 늘려주는 코드 종료
pd.reset_option('display.max_rows', None)

# 결측치 제거 후 인덱스 초기화
kosdaq_na.reset_index(drop=True, inplace=True)

kosdaq_na.to_csv('./datasets/2.kosdaq_결측치제거완료.csv')
