==> 4. kosdaq_더미화.ipynb <==
import pandas as pd

kosdaq = pd.read_csv('./datasets/3.kosdaq_이상치처리완료.csv', index_col=0)
kosdaq

# 산업분류 더미화

kosdaq['통계청 한국표준산업분류 10차(대분류)'].value_counts()

def change(x):
    if x == '제조업':
        return 1
    else:
        return 0
    
kosdaq['산업분류'] = kosdaq['통계청 한국표준산업분류 10차(대분류)'].apply(change)

kosdaq['산업분류'].value_counts()

kosdaq.drop(columns='통계청 한국표준산업분류 10차(대분류)', axis=1, inplace=True)

# 기업생애주기 더미화

dummies_genre = pd.get_dummies(kosdaq['기업생애주기'], prefix='기업생애주기', dtype=int)
df = pd.concat([kosdaq,dummies_genre], axis=1)

df.drop(columns='기업생애주기', axis=1, inplace=True)

df

df['label'].value_counts()

df.to_csv('./datasets/4.kosdaq_더미화완료.csv')
