==> 4. kospi_더미화.ipynb <==
import pandas as pd

kospi = pd.read_csv('./datasets/3.kospi_이상치처리완료.csv', index_col=0)
kospi

# 산업분류 더미화

kospi['통계청 한국표준산업분류 10차(대분류)'].value_counts()

def change(x):
    if x == '제조업':
        return 1
    else:
        return 0
    
kospi['산업분류'] = kospi['통계청 한국표준산업분류 10차(대분류)'].apply(change)

kospi['산업분류'].value_counts()

kospi.drop(columns='통계청 한국표준산업분류 10차(대분류)', axis=1, inplace=True)

# 기업생애주기 더미화

dummies_genre = pd.get_dummies(kospi['기업생애주기'], prefix='기업생애주기', dtype=int)
df = pd.concat([kospi,dummies_genre], axis=1)

df.drop(columns='기업생애주기', axis=1, inplace=True)

df

df['label'].value_counts()

df.to_csv('./datasets/4.kospi_더미화완료.csv')
