==> 5. kosdaq+kospi_데이터결합+최종데이터.ipynb <==
import pandas as pd

kospi = pd.read_csv('./datasets/4.kospi_더미화완료.csv', index_col=0)
kosdaq = pd.read_csv('./datasets/4.kosdaq_더미화완료.csv', index_col=0)

kospi

kosdaq

# 데이터 결합

df_concat = pd.concat([kospi,kosdaq], axis=0)

df_concat.reset_index(drop=True,inplace=True)

df_concat.info()

# 결합과정 중 중복데이터 제거 코드

df_concat.drop_duplicates(subset=['거래소코드', '회계년도'], keep='first', inplace=True)

df_concat.reset_index(drop=True,inplace=True)

df_concat.to_csv('./datasets/KOSDAQ+KOSPI_DATA.csv')
kosdaq.to_csv('./datasets/KOSDAQ_DATA.csv')

# Kospi Data Spilt

# 회계년도 기준 데이터 스플릿 진행
df_concat_1117 = df_concat[df_concat['회계년도'] <= '2017/12']
df_concat_1819 = df_concat[(df_concat['회계년도'] <= '2019/12') & (df_concat['회계년도'] >= '2018/12')]
df_concat_2021 = df_concat[(df_concat['회계년도'] <= '2021/12') & (df_concat['회계년도'] >= '2020/12')]

df_concat_1117['label'].value_counts()

df_concat_1819['label'].value_counts()

# 인덱스 초기화
df_concat_1117.reset_index(drop=True, inplace=True)
df_concat_1819.reset_index(drop=True, inplace=True)
df_concat_2021.reset_index(drop=True, inplace=True)

df_concat_1117['label'].value_counts()

df_concat_1819['label'].value_counts()

# Train, Test1, Test2 Data 저장
df_concat_1117.to_csv('./datasets/KOSDAQ+KOSPI_TRAIN_DATA.csv')
df_concat_1819.to_csv('./datasets/KOSDAQ+KOSPI_TEST1_DATA.csv')
df_concat_2021.to_csv('./datasets/KOSDAQ+KOSPI_TEST2_DATA.csv')

# Kosdaq Data Spilt

# 회계년도 기준 데이터 스플릿 진행
kosdaq_1117 = kosdaq[kosdaq['회계년도'] <= '2017/12']
kosdaq_1819 = kosdaq[(kosdaq['회계년도'] <= '2019/12') & (kosdaq['회계년도'] >= '2018/12')]
kosdaq_2021 = kosdaq[(kosdaq['회계년도'] <= '2021/12') & (kosdaq['회계년도'] >= '2020/12')]

kosdaq_1117['label'].value_counts()

kosdaq_1819['label'].value_counts()

# 인덱스 초기화
kosdaq_1117.reset_index(drop=True, inplace=True)
kosdaq_1819.reset_index(drop=True, inplace=True)
kosdaq_2021.reset_index(drop=True, inplace=True)

# Train, Test1, Test2 Data 저장
kosdaq_1117.to_csv('./datasets/KOSDAQ_TRAIN_DATA.csv')
kosdaq_1819.to_csv('./datasets/KOSDAQ_TEST1_DATA.csv')
kosdaq_2021.to_csv('./datasets/KOSDAQ_TEST2_DATA.csv')
