==> 8. DS2(코스피+코스닥)_교차검증.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
import seaborn as sns
from datetime import datetime
import time
import warnings
warnings.filterwarnings("ignore")

from sklearn.preprocessing import MinMaxScaler, StandardScaler, RobustScaler
from sklearn.metrics import accuracy_score,f1_score,recall_score,precision_score,confusion_matrix,ConfusionMatrixDisplay,roc_curve,roc_auc_score,precision_recall_curve
from sklearn.ensemble import RandomForestClassifier , StackingClassifier
from sklearn.svm import SVC
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.tree import DecisionTreeClassifier
from xgboost import XGBClassifier
from lightgbm import LGBMClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression,Lasso
from sklearn.preprocessing import Binarizer
from sklearn.model_selection import cross_val_score,GridSearchCV

# 한글 깨짐 방지
import matplotlib.pyplot as plt
plt.rcParams['font.family'] = 'Malgun Gothic'

kk = pd.read_csv('./datasets/KOSDAQ+KOSPI_DATA.csv', index_col=0)
kk = kk.iloc[:,2:]
kosdaq = pd.read_csv('./datasets/KOSDAQ_DATA.csv', index_col=0)
kosdaq = kosdaq.iloc[:,2:]

kk[kk['회계년도'] == '2011/12']['label'].value_counts()

kk[kk['회계년도'] == '2012/12']['label'].value_counts()

kk[kk['회계년도'] == '2013/12']['label'].value_counts()

kk[kk['회계년도'] == '2014/12']['label'].value_counts()

kk[kk['회계년도'] == '2015/12']['label'].value_counts()

kk[kk['회계년도'] == '2016/12']['label'].value_counts()

kk[kk['회계년도'] == '2017/12']['label'].value_counts()

kk[kk['회계년도'] == '2018/12']['label'].value_counts()

kk[kk['회계년도'] == '2019/12']['label'].value_counts()

train_df = pd.read_csv('./datasets/KOSDAQ+KOSPI_TRAIN_DATA.csv', index_col=0)
train_df = train_df.iloc[:,2:]
test_df = pd.read_csv('./datasets/KOSDAQ+KOSPI_TEST1_DATA.csv', index_col=0)
test_df = test_df.iloc[:,2:]

# 코스닥 + 코스피

# 교차검증 1 데이터 셋
kk_1112 = kk[(kk['회계년도'] <= '2012/12') & (kk['회계년도'] >= '2011/12')]
kk_1319 = kk[(kk['회계년도'] <= '2019/12') & (kk['회계년도'] >= '2013/12')]
# 교차검증 2 데이터 셋
kk_1213 = kk[(kk['회계년도'] <= '2013/12') & (kk['회계년도'] >= '2012/12')]
kk_111419 = kk[(kk['회계년도'] <= '2019/12') & (kk['회계년도'] >= '2014/12') | (kk['회계년도'] == '2011/12')]
# 교차검증 3 데이터 셋
kk_1314 = kk[(kk['회계년도'] <= '2014/12') & (kk['회계년도'] >= '2013/12')]
kk_11121519 = kk[(kk['회계년도'] <= '2019/12') & (kk['회계년도'] >= '2015/12') | (kk['회계년도'] >= '2011/12') & (kk['회계년도'] <= '2012/12')]
# 교차검증 4 데이터 셋
kk_1415 = kk[(kk['회계년도'] <= '2015/12') & (kk['회계년도'] >= '2014/12')]
kk_11131619 = kk[(kk['회계년도'] <= '2019/12') & (kk['회계년도'] >= '2016/12') | (kk['회계년도'] >= '2011/12') & (kk['회계년도'] <= '2013/12')]
# 교차검증 5 데이터 셋
kk_1516 = kk[(kk['회계년도'] <= '2016/12') & (kk['회계년도'] >= '2015/12')]
kk_11141719 = kk[(kk['회계년도'] <= '2019/12') & (kk['회계년도'] >= '2017/12') | (kk['회계년도'] >= '2011/12') & (kk['회계년도'] <= '2014/12')]
# 교차검증 6 데이터 셋
kk_1617 = kk[(kk['회계년도'] <= '2017/12') & (kk['회계년도'] >= '2016/12')]
kk_11151819 = kk[(kk['회계년도'] <= '2019/12') & (kk['회계년도'] >= '2018/12') | (kk['회계년도'] >= '2011/12') & (kk['회계년도'] <= '2015/12')]
# 교차검증 7 데이터 셋
kk_1718 = kk[(kk['회계년도'] <= '2018/12') & (kk['회계년도'] >= '2017/12')]
kk_111619 = kk[(kk['회계년도'] == '2019/12') | (kk['회계년도'] >= '2011/12') & (kk['회계년도'] <= '2016/12')]

kk_1718['label'].value_counts()

# 코스닥

# 교차검증 1 데이터 셋
kosdaq_1112 = kosdaq[(kosdaq['회계년도'] <= '2012/12') & (kosdaq['회계년도'] >= '2011/12')]
kosdaq_1319 = kosdaq[(kosdaq['회계년도'] <= '2019/12') & (kosdaq['회계년도'] >= '2013/12')]
# 교차검증 2 데이터 셋
kosdaq_1213 = kosdaq[(kosdaq['회계년도'] <= '2013/12') & (kosdaq['회계년도'] >= '2012/12')]
kosdaq_111419 = kosdaq[(kosdaq['회계년도'] <= '2019/12') & (kosdaq['회계년도'] >= '2014/12') | (kosdaq['회계년도'] == '2011/12')]
# 교차검증 3 데이터 셋
kosdaq_1314 = kosdaq[(kosdaq['회계년도'] <= '2014/12') & (kosdaq['회계년도'] >= '2013/12')]
kosdaq_11121519 = kosdaq[(kosdaq['회계년도'] <= '2019/12') & (kosdaq['회계년도'] >= '2015/12') | (kosdaq['회계년도'] >= '2011/12') & (kosdaq['회계년도'] <= '2012/12')]
# 교차검증 4 데이터 셋
kosdaq_1415 = kosdaq[(kosdaq['회계년도'] <= '2015/12') & (kosdaq['회계년도'] >= '2014/12')]
kosdaq_11131619 = kosdaq[(kosdaq['회계년도'] <= '2019/12') & (kosdaq['회계년도'] >= '2016/12') | (kosdaq['회계년도'] >= '2011/12') & (kosdaq['회계년도'] <= '2013/12')]
# 교차검증 5 데이터 셋
kosdaq_1516 = kosdaq[(kosdaq['회계년도'] <= '2016/12') & (kosdaq['회계년도'] >= '2015/12')]
kosdaq_11141719 = kosdaq[(kosdaq['회계년도'] <= '2019/12') & (kosdaq['회계년도'] >= '2017/12') | (kosdaq['회계년도'] >= '2011/12') & (kosdaq['회계년도'] <= '2014/12')]
# 교차검증 6 데이터 셋
kosdaq_1617 = kosdaq[(kosdaq['회계년도'] <= '2017/12') & (kosdaq['회계년도'] >= '2016/12')]
kosdaq_11151819 = kosdaq[(kosdaq['회계년도'] <= '2019/12') & (kosdaq['회계년도'] >= '2018/12') | (kosdaq['회계년도'] >= '2011/12') & (kosdaq['회계년도'] <= '2015/12')]
# 교차검증 7 데이터 셋
kosdaq_1718 = kosdaq[(kosdaq['회계년도'] <= '2018/12') & (kosdaq['회계년도'] >= '2017/12')]
kosdaq_111619 = kosdaq[(kosdaq['회계년도'] == '2019/12') | (kosdaq['회계년도'] >= '2011/12') & (kosdaq['회계년도'] <= '2016/12')]

kk[(kk['회계년도'] <= '2012/12') & (kk['회계년도'] >= '2011/12')]

kk[(kk['회계년도'] <= '2019/12') & (kk['회계년도'] >= '2013/12')]['label'].value_counts()

# 검증 함수 만들기

# 1번 데이터 검증

def model1(Data):

    train = Data[(Data['회계년도'] <= '2019/12') & (Data['회계년도'] >= '2013/12')].reset_index(drop=True)
    test = Data[(Data['회계년도'] <= '2012/12') & (Data['회계년도'] >= '2011/12')].reset_index(drop=True)

    ## Train Scaling 전 Data 분리하기 ##

    X_train = train.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_train_concat = train[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_train = train[['label']]
    X_test = test.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_test_concat = test[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_test = test[['label']]


    ## Scaling##

    from sklearn.preprocessing import StandardScaler
    # StandardScaler() Scaler객체 생성.
    scaler = StandardScaler()
    # 학습 데이터에 대해서 fit(), transform() 수행.
    scaler.fit(X_train)
    train_scaled_SS = scaler.transform(X_train)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_train_SS = pd.DataFrame(data = train_scaled_SS, columns=X_train.columns)

    # 테스트 데이터에서는 다시 fit(), transform()이나 fit_transform()을 수행하지 않고 transform만 수행.
    test_scaled_MM = scaler.transform(X_test)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_test_SS = pd.DataFrame(data = test_scaled_MM, columns=X_test.columns)


    X_train_SS = pd.concat([X_train_SS,X_train_concat,y_train], axis=1)
    X_test_SS = pd.concat([X_test_SS,X_test_concat, y_test], axis=1)

    trian_test_df = pd.concat([X_train_SS, X_test_SS], axis=0)
    trian_test_df.reset_index(drop=True, inplace=True)      

    df_2011 = trian_test_df[trian_test_df['회계년도'] == '2011/12']
    df_2012 = trian_test_df[trian_test_df['회계년도'] == '2012/12']
    df_2013 = trian_test_df[trian_test_df['회계년도'] == '2013/12']
    df_2014 = trian_test_df[trian_test_df['회계년도'] == '2014/12']
    df_2015 = trian_test_df[trian_test_df['회계년도'] == '2015/12']
    df_2016 = trian_test_df[trian_test_df['회계년도'] == '2016/12']
    df_2017 = trian_test_df[trian_test_df['회계년도'] == '2017/12']
    df_2018 = trian_test_df[trian_test_df['회계년도'] == '2018/12']
    df_2019 = trian_test_df[trian_test_df['회계년도'] == '2019/12']

    df_X_2011 = df_2011.drop(columns='label', axis=1)
    df_y_2011 = df_2011[['label']]
    df_X_2012 = df_2012.drop(columns='label', axis=1)
    df_y_2012 = df_2012[['label']]
    df_X_2013 = df_2013.drop(columns='label', axis=1)
    df_y_2013 = df_2013[['label']]
    df_X_2014 = df_2014.drop(columns='label', axis=1)
    df_y_2014 = df_2014[['label']]
    df_X_2015 = df_2015.drop(columns='label', axis=1)
    df_y_2015 = df_2015[['label']]
    df_X_2016 = df_2016.drop(columns='label', axis=1)
    df_y_2016 = df_2016[['label']]
    df_X_2017 = df_2017.drop(columns='label', axis=1)
    df_y_2017 = df_2017[['label']]
    df_X_2018 = df_2018.drop(columns='label', axis=1)
    df_y_2018 = df_2018[['label']]
    df_X_2019 = df_2019.drop(columns='label', axis=1)
    df_y_2019 = df_2019[['label']]
    

    ## UnderSampler ##

    from imblearn.under_sampling import RandomUnderSampler

    X_samp_11, y_samp_11 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2011, df_y_2011)
    X_samp_12, y_samp_12 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2012, df_y_2012)
    X_samp_13, y_samp_13 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2013, df_y_2013)
    X_samp_14, y_samp_14 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2014, df_y_2014)
    X_samp_15, y_samp_15 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2015, df_y_2015)
    X_samp_16, y_samp_16 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2016, df_y_2016)
    X_samp_17, y_samp_17 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2017, df_y_2017)
    X_samp_18, y_samp_18 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2018, df_y_2018)
    X_samp_19, y_samp_19 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2019, df_y_2019)

    X_samp = pd.concat([X_samp_11, X_samp_12, X_samp_13, X_samp_14, X_samp_15, X_samp_16, X_samp_17, X_samp_18, X_samp_19], axis=0)
    y_samp = pd.concat([y_samp_11, y_samp_12, y_samp_13, y_samp_14, y_samp_15, y_samp_16, y_samp_17, y_samp_18, y_samp_19], axis=0)

    trian_test_samp_df = pd.concat([X_samp, y_samp], axis=1)

    # 교차검증 1 데이터 셋
    X_samp_train = trian_test_samp_df[(trian_test_samp_df['회계년도'] <= '2019/12') & (trian_test_samp_df['회계년도'] >= '2013/12')]
    X_samp_test = trian_test_samp_df[(trian_test_samp_df['회계년도'] <= '2012/12') & (trian_test_samp_df['회계년도'] >= '2011/12')]

    X_samp_train_1 = X_samp_train.drop(columns=['label', '회계년도'], axis=1)
    y_samp_train_1 = X_samp_train[['label']]

    X_samp_test_1 = X_samp_test.drop(columns=['label', '회계년도'], axis=1)
    y_samp_test_1 = X_samp_test[['label']]

    X_samp_train_1.reset_index(drop=True, inplace=True)
    X_samp_test_1.reset_index(drop=True, inplace=True)
    y_samp_train_1.reset_index(drop=True, inplace=True)
    y_samp_test_1.reset_index(drop=True, inplace=True)

    from imblearn.over_sampling import BorderlineSMOTE

    X_samp_2, y_samp_2 = BorderlineSMOTE(random_state=0,sampling_strategy=1).fit_resample(X_samp_train_1, y_samp_train_1)

    train_df = pd.concat([X_samp_2, y_samp_2], axis=1)
    test_df = pd.concat([X_samp_test_1, y_samp_test_1], axis=1)

    return train_df, test_df

model1(kk)[0]['label'].value_counts()

model1(kk)[1]['label'].value_counts()

# 2번 데이터 검증

def model2(Data):

    train = Data[(Data['회계년도'] <= '2019/12') & (Data['회계년도'] >= '2014/12') | (Data['회계년도'] == '2011/12')].reset_index(drop=True)
    test = Data[(Data['회계년도'] <= '2013/12') & (Data['회계년도'] >= '2012/12')].reset_index(drop=True)
    ## Train Scaling 전 Data 분리하기 ##

    X_train = train.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_train_concat = train[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_train = train[['label']]
    X_test = test.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_test_concat = test[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_test = test[['label']]


    ## Scaling##

    from sklearn.preprocessing import StandardScaler
    # StandardScaler() Scaler객체 생성.
    scaler = StandardScaler()
    # 학습 데이터에 대해서 fit(), transform() 수행.
    scaler.fit(X_train)
    train_scaled_SS = scaler.transform(X_train)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_train_SS = pd.DataFrame(data = train_scaled_SS, columns=X_train.columns)

    # 테스트 데이터에서는 다시 fit(), transform()이나 fit_transform()을 수행하지 않고 transform만 수행.
    test_scaled_MM = scaler.transform(X_test)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_test_SS = pd.DataFrame(data = test_scaled_MM, columns=X_test.columns)


    X_train_SS = pd.concat([X_train_SS,X_train_concat,y_train], axis=1)
    X_test_SS = pd.concat([X_test_SS,X_test_concat, y_test], axis=1)

    trian_test_df = pd.concat([X_train_SS, X_test_SS], axis=0)
    trian_test_df.reset_index(drop=True, inplace=True)      

    df_2011 = trian_test_df[trian_test_df['회계년도'] == '2011/12']
    df_2012 = trian_test_df[trian_test_df['회계년도'] == '2012/12']
    df_2013 = trian_test_df[trian_test_df['회계년도'] == '2013/12']
    df_2014 = trian_test_df[trian_test_df['회계년도'] == '2014/12']
    df_2015 = trian_test_df[trian_test_df['회계년도'] == '2015/12']
    df_2016 = trian_test_df[trian_test_df['회계년도'] == '2016/12']
    df_2017 = trian_test_df[trian_test_df['회계년도'] == '2017/12']
    df_2018 = trian_test_df[trian_test_df['회계년도'] == '2018/12']
    df_2019 = trian_test_df[trian_test_df['회계년도'] == '2019/12']

    df_X_2011 = df_2011.drop(columns='label', axis=1)
    df_y_2011 = df_2011[['label']]
    df_X_2012 = df_2012.drop(columns='label', axis=1)
    df_y_2012 = df_2012[['label']]
    df_X_2013 = df_2013.drop(columns='label', axis=1)
    df_y_2013 = df_2013[['label']]
    df_X_2014 = df_2014.drop(columns='label', axis=1)
    df_y_2014 = df_2014[['label']]
    df_X_2015 = df_2015.drop(columns='label', axis=1)
    df_y_2015 = df_2015[['label']]
    df_X_2016 = df_2016.drop(columns='label', axis=1)
    df_y_2016 = df_2016[['label']]
    df_X_2017 = df_2017.drop(columns='label', axis=1)
    df_y_2017 = df_2017[['label']]
    df_X_2018 = df_2018.drop(columns='label', axis=1)
    df_y_2018 = df_2018[['label']]
    df_X_2019 = df_2019.drop(columns='label', axis=1)
    df_y_2019 = df_2019[['label']]
    

    ## UnderSampler ##

    from imblearn.under_sampling import RandomUnderSampler

    X_samp_11, y_samp_11 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2011, df_y_2011)
    X_samp_12, y_samp_12 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2012, df_y_2012)
    X_samp_13, y_samp_13 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2013, df_y_2013)
    X_samp_14, y_samp_14 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2014, df_y_2014)
    X_samp_15, y_samp_15 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2015, df_y_2015)
    X_samp_16, y_samp_16 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2016, df_y_2016)
    X_samp_17, y_samp_17 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2017, df_y_2017)
    X_samp_18, y_samp_18 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2018, df_y_2018)
    X_samp_19, y_samp_19 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2019, df_y_2019)

    X_samp = pd.concat([X_samp_11, X_samp_12, X_samp_13, X_samp_14, X_samp_15, X_samp_16, X_samp_17, X_samp_18, X_samp_19], axis=0)
    y_samp = pd.concat([y_samp_11, y_samp_12, y_samp_13, y_samp_14, y_samp_15, y_samp_16, y_samp_17, y_samp_18, y_samp_19], axis=0)

    a = pd.concat([X_samp, y_samp], axis=1)

    # 교차검증 2 데이터 셋
    X_samp_train = a[(a['회계년도'] <= '2019/12') & (a['회계년도'] >= '2014/12') | (a['회계년도'] == '2011/12')]
    X_samp_test = a[(a['회계년도'] <= '2013/12') & (a['회계년도'] >= '2012/12')]

    X_samp_train_1 = X_samp_train.drop(columns=['label', '회계년도'], axis=1)
    y_samp_train_1 = X_samp_train[['label']]

    X_samp_test_1 = X_samp_test.drop(columns=['label', '회계년도'], axis=1)
    y_samp_test_1 = X_samp_test[['label']]

    X_samp_train_1.reset_index(drop=True, inplace=True)
    X_samp_test_1.reset_index(drop=True, inplace=True)
    y_samp_train_1.reset_index(drop=True, inplace=True)
    y_samp_test_1.reset_index(drop=True, inplace=True)

    from imblearn.over_sampling import BorderlineSMOTE

    X_samp_2, y_samp_2 = BorderlineSMOTE(random_state=0,sampling_strategy=1).fit_resample(X_samp_train_1, y_samp_train_1)

    train_df = pd.concat([X_samp_2, y_samp_2], axis=1)
    test_df = pd.concat([X_samp_test_1, y_samp_test_1], axis=1)

    return train_df, test_df

model2(kk)[0]['label'].value_counts()

model2(kk)[1]['label'].value_counts()

# 3번 데이터 검증

def model3(Data):

    train = Data[(Data['회계년도'] <= '2019/12') & (Data['회계년도'] >= '2015/12') | (Data['회계년도'] >= '2011/12') & (Data['회계년도'] <= '2012/12')].reset_index(drop=True)
    test = Data[(Data['회계년도'] <= '2014/12') & (Data['회계년도'] >= '2013/12')].reset_index(drop=True)

    X_train = train.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_train_concat = train[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_train = train[['label']]
    X_test = test.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_test_concat = test[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_test = test[['label']]


    ## Scaling##

    from sklearn.preprocessing import StandardScaler
    # StandardScaler() Scaler객체 생성.
    scaler = StandardScaler()
    # 학습 데이터에 대해서 fit(), transform() 수행.
    scaler.fit(X_train)
    train_scaled_SS = scaler.transform(X_train)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_train_SS = pd.DataFrame(data = train_scaled_SS, columns=X_train.columns)

    # 테스트 데이터에서는 다시 fit(), transform()이나 fit_transform()을 수행하지 않고 transform만 수행.
    test_scaled_MM = scaler.transform(X_test)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_test_SS = pd.DataFrame(data = test_scaled_MM, columns=X_test.columns)


    X_train_SS = pd.concat([X_train_SS,X_train_concat,y_train], axis=1)
    X_test_SS = pd.concat([X_test_SS,X_test_concat, y_test], axis=1)

    trian_test_df = pd.concat([X_train_SS, X_test_SS], axis=0)
    trian_test_df.reset_index(drop=True, inplace=True)      

    df_2011 = trian_test_df[trian_test_df['회계년도'] == '2011/12']
    df_2012 = trian_test_df[trian_test_df['회계년도'] == '2012/12']
    df_2013 = trian_test_df[trian_test_df['회계년도'] == '2013/12']
    df_2014 = trian_test_df[trian_test_df['회계년도'] == '2014/12']
    df_2015 = trian_test_df[trian_test_df['회계년도'] == '2015/12']
    df_2016 = trian_test_df[trian_test_df['회계년도'] == '2016/12']
    df_2017 = trian_test_df[trian_test_df['회계년도'] == '2017/12']
    df_2018 = trian_test_df[trian_test_df['회계년도'] == '2018/12']
    df_2019 = trian_test_df[trian_test_df['회계년도'] == '2019/12']

    df_X_2011 = df_2011.drop(columns='label', axis=1)
    df_y_2011 = df_2011[['label']]
    df_X_2012 = df_2012.drop(columns='label', axis=1)
    df_y_2012 = df_2012[['label']]
    df_X_2013 = df_2013.drop(columns='label', axis=1)
    df_y_2013 = df_2013[['label']]
    df_X_2014 = df_2014.drop(columns='label', axis=1)
    df_y_2014 = df_2014[['label']]
    df_X_2015 = df_2015.drop(columns='label', axis=1)
    df_y_2015 = df_2015[['label']]
    df_X_2016 = df_2016.drop(columns='label', axis=1)
    df_y_2016 = df_2016[['label']]
    df_X_2017 = df_2017.drop(columns='label', axis=1)
    df_y_2017 = df_2017[['label']]
    df_X_2018 = df_2018.drop(columns='label', axis=1)
    df_y_2018 = df_2018[['label']]
    df_X_2019 = df_2019.drop(columns='label', axis=1)
    df_y_2019 = df_2019[['label']]
    

    ## UnderSampler ##

    from imblearn.under_sampling import RandomUnderSampler

    X_samp_11, y_samp_11 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2011, df_y_2011)
    X_samp_12, y_samp_12 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2012, df_y_2012)
    X_samp_13, y_samp_13 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2013, df_y_2013)
    X_samp_14, y_samp_14 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2014, df_y_2014)
    X_samp_15, y_samp_15 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2015, df_y_2015)
    X_samp_16, y_samp_16 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2016, df_y_2016)
    X_samp_17, y_samp_17 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2017, df_y_2017)
    X_samp_18, y_samp_18 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2018, df_y_2018)
    X_samp_19, y_samp_19 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2019, df_y_2019)

    X_samp = pd.concat([X_samp_11, X_samp_12, X_samp_13, X_samp_14, X_samp_15, X_samp_16, X_samp_17, X_samp_18, X_samp_19], axis=0)
    y_samp = pd.concat([y_samp_11, y_samp_12, y_samp_13, y_samp_14, y_samp_15, y_samp_16, y_samp_17, y_samp_18, y_samp_19], axis=0)

    a = pd.concat([X_samp, y_samp], axis=1)

    # 교차검증 3 데이터 셋
    X_samp_train = a[(a['회계년도'] <= '2019/12') & (a['회계년도'] >= '2015/12') | (a['회계년도'] >= '2011/12') & (a['회계년도'] <= '2012/12')]
    X_samp_test = a[(a['회계년도'] <= '2014/12') & (a['회계년도'] >= '2013/12')]

    X_samp_train_1 = X_samp_train.drop(columns=['label', '회계년도'], axis=1)
    y_samp_train_1 = X_samp_train[['label']]

    X_samp_test_1 = X_samp_test.drop(columns=['label', '회계년도'], axis=1)
    y_samp_test_1 = X_samp_test[['label']]

    X_samp_train_1.reset_index(drop=True, inplace=True)
    X_samp_test_1.reset_index(drop=True, inplace=True)
    y_samp_train_1.reset_index(drop=True, inplace=True)
    y_samp_test_1.reset_index(drop=True, inplace=True)

    from imblearn.over_sampling import BorderlineSMOTE

    X_samp_2, y_samp_2 = BorderlineSMOTE(random_state=0,sampling_strategy=1).fit_resample(X_samp_train_1, y_samp_train_1)

    train_df = pd.concat([X_samp_2, y_samp_2], axis=1)
    test_df = pd.concat([X_samp_test_1, y_samp_test_1], axis=1)

    return train_df, test_df

model3(kk)[0]['label'].value_counts()

model3(kk)[1]['label'].value_counts()

# 4번 데이터 검증

def model4(Data):

    train = Data[(Data['회계년도'] <= '2019/12') & (Data['회계년도'] >= '2016/12') | (Data['회계년도'] >= '2011/12') & (Data['회계년도'] <= '2013/12')].reset_index(drop=True)
    test = Data[(Data['회계년도'] <= '2015/12') & (Data['회계년도'] >= '2014/12')].reset_index(drop=True)

    X_train = train.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_train_concat = train[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_train = train[['label']]
    X_test = test.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_test_concat = test[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_test = test[['label']]


    ## Scaling##

    from sklearn.preprocessing import StandardScaler
    # StandardScaler() Scaler객체 생성.
    scaler = StandardScaler()
    # 학습 데이터에 대해서 fit(), transform() 수행.
    scaler.fit(X_train)
    train_scaled_SS = scaler.transform(X_train)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_train_SS = pd.DataFrame(data = train_scaled_SS, columns=X_train.columns)

    # 테스트 데이터에서는 다시 fit(), transform()이나 fit_transform()을 수행하지 않고 transform만 수행.
    test_scaled_MM = scaler.transform(X_test)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_test_SS = pd.DataFrame(data = test_scaled_MM, columns=X_test.columns)


    X_train_SS = pd.concat([X_train_SS,X_train_concat,y_train], axis=1)
    X_test_SS = pd.concat([X_test_SS,X_test_concat, y_test], axis=1)

    trian_test_df = pd.concat([X_train_SS, X_test_SS], axis=0)
    trian_test_df.reset_index(drop=True, inplace=True)      

    df_2011 = trian_test_df[trian_test_df['회계년도'] == '2011/12']
    df_2012 = trian_test_df[trian_test_df['회계년도'] == '2012/12']
    df_2013 = trian_test_df[trian_test_df['회계년도'] == '2013/12']
    df_2014 = trian_test_df[trian_test_df['회계년도'] == '2014/12']
    df_2015 = trian_test_df[trian_test_df['회계년도'] == '2015/12']
    df_2016 = trian_test_df[trian_test_df['회계년도'] == '2016/12']
    df_2017 = trian_test_df[trian_test_df['회계년도'] == '2017/12']
    df_2018 = trian_test_df[trian_test_df['회계년도'] == '2018/12']
    df_2019 = trian_test_df[trian_test_df['회계년도'] == '2019/12']

    df_X_2011 = df_2011.drop(columns='label', axis=1)
    df_y_2011 = df_2011[['label']]
    df_X_2012 = df_2012.drop(columns='label', axis=1)
    df_y_2012 = df_2012[['label']]
    df_X_2013 = df_2013.drop(columns='label', axis=1)
    df_y_2013 = df_2013[['label']]
    df_X_2014 = df_2014.drop(columns='label', axis=1)
    df_y_2014 = df_2014[['label']]
    df_X_2015 = df_2015.drop(columns='label', axis=1)
    df_y_2015 = df_2015[['label']]
    df_X_2016 = df_2016.drop(columns='label', axis=1)
    df_y_2016 = df_2016[['label']]
    df_X_2017 = df_2017.drop(columns='label', axis=1)
    df_y_2017 = df_2017[['label']]
    df_X_2018 = df_2018.drop(columns='label', axis=1)
    df_y_2018 = df_2018[['label']]
    df_X_2019 = df_2019.drop(columns='label', axis=1)
    df_y_2019 = df_2019[['label']]
    

    ## UnderSampler ##

    from imblearn.under_sampling import RandomUnderSampler

    X_samp_11, y_samp_11 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2011, df_y_2011)
    X_samp_12, y_samp_12 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2012, df_y_2012)
    X_samp_13, y_samp_13 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2013, df_y_2013)
    X_samp_14, y_samp_14 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2014, df_y_2014)
    X_samp_15, y_samp_15 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2015, df_y_2015)
    X_samp_16, y_samp_16 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2016, df_y_2016)
    X_samp_17, y_samp_17 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2017, df_y_2017)
    X_samp_18, y_samp_18 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2018, df_y_2018)
    X_samp_19, y_samp_19 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2019, df_y_2019)

    X_samp = pd.concat([X_samp_11, X_samp_12, X_samp_13, X_samp_14, X_samp_15, X_samp_16, X_samp_17, X_samp_18, X_samp_19], axis=0)
    y_samp = pd.concat([y_samp_11, y_samp_12, y_samp_13, y_samp_14, y_samp_15, y_samp_16, y_samp_17, y_samp_18, y_samp_19], axis=0)

    a = pd.concat([X_samp, y_samp], axis=1)

    # 교차검증 3 데이터 셋
    X_samp_train = a[(a['회계년도'] <= '2019/12') & (a['회계년도'] >= '2016/12') | (a['회계년도'] >= '2011/12') & (a['회계년도'] <= '2013/12')]
    X_samp_test = a[(a['회계년도'] <= '2015/12') & (a['회계년도'] >= '2014/12')]

    X_samp_train_1 = X_samp_train.drop(columns=['label', '회계년도'], axis=1)
    y_samp_train_1 = X_samp_train[['label']]

    X_samp_test_1 = X_samp_test.drop(columns=['label', '회계년도'], axis=1)
    y_samp_test_1 = X_samp_test[['label']]

    X_samp_train_1.reset_index(drop=True, inplace=True)
    X_samp_test_1.reset_index(drop=True, inplace=True)
    y_samp_train_1.reset_index(drop=True, inplace=True)
    y_samp_test_1.reset_index(drop=True, inplace=True)

    from imblearn.over_sampling import BorderlineSMOTE

    X_samp_2, y_samp_2 = BorderlineSMOTE(random_state=0,sampling_strategy=1).fit_resample(X_samp_train_1, y_samp_train_1)

    train_df = pd.concat([X_samp_2, y_samp_2], axis=1)
    test_df = pd.concat([X_samp_test_1, y_samp_test_1], axis=1)

    return train_df, test_df

model4(kk)[0]['label'].value_counts()

model4(kk)[1]['label'].value_counts()

# 5번 데이터 검증

def model5(Data):

    train = Data[(Data['회계년도'] <= '2019/12') & (Data['회계년도'] >= '2017/12') | (Data['회계년도'] >= '2011/12') & (Data['회계년도'] <= '2014/12')].reset_index(drop=True)
    test = Data[(Data['회계년도'] <= '2016/12') & (Data['회계년도'] >= '2015/12')].reset_index(drop=True)

    X_train = train.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_train_concat = train[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_train = train[['label']]
    X_test = test.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_test_concat = test[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_test = test[['label']]


    ## Scaling##

    from sklearn.preprocessing import StandardScaler
    # StandardScaler() Scaler객체 생성.
    scaler = StandardScaler()
    # 학습 데이터에 대해서 fit(), transform() 수행.
    scaler.fit(X_train)
    train_scaled_SS = scaler.transform(X_train)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_train_SS = pd.DataFrame(data = train_scaled_SS, columns=X_train.columns)

    # 테스트 데이터에서는 다시 fit(), transform()이나 fit_transform()을 수행하지 않고 transform만 수행.
    test_scaled_MM = scaler.transform(X_test)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_test_SS = pd.DataFrame(data = test_scaled_MM, columns=X_test.columns)


    X_train_SS = pd.concat([X_train_SS,X_train_concat,y_train], axis=1)
    X_test_SS = pd.concat([X_test_SS,X_test_concat, y_test], axis=1)

    trian_test_df = pd.concat([X_train_SS, X_test_SS], axis=0)
    trian_test_df.reset_index(drop=True, inplace=True)      

    df_2011 = trian_test_df[trian_test_df['회계년도'] == '2011/12']
    df_2012 = trian_test_df[trian_test_df['회계년도'] == '2012/12']
    df_2013 = trian_test_df[trian_test_df['회계년도'] == '2013/12']
    df_2014 = trian_test_df[trian_test_df['회계년도'] == '2014/12']
    df_2015 = trian_test_df[trian_test_df['회계년도'] == '2015/12']
    df_2016 = trian_test_df[trian_test_df['회계년도'] == '2016/12']
    df_2017 = trian_test_df[trian_test_df['회계년도'] == '2017/12']
    df_2018 = trian_test_df[trian_test_df['회계년도'] == '2018/12']
    df_2019 = trian_test_df[trian_test_df['회계년도'] == '2019/12']

    df_X_2011 = df_2011.drop(columns='label', axis=1)
    df_y_2011 = df_2011[['label']]
    df_X_2012 = df_2012.drop(columns='label', axis=1)
    df_y_2012 = df_2012[['label']]
    df_X_2013 = df_2013.drop(columns='label', axis=1)
    df_y_2013 = df_2013[['label']]
    df_X_2014 = df_2014.drop(columns='label', axis=1)
    df_y_2014 = df_2014[['label']]
    df_X_2015 = df_2015.drop(columns='label', axis=1)
    df_y_2015 = df_2015[['label']]
    df_X_2016 = df_2016.drop(columns='label', axis=1)
    df_y_2016 = df_2016[['label']]
    df_X_2017 = df_2017.drop(columns='label', axis=1)
    df_y_2017 = df_2017[['label']]
    df_X_2018 = df_2018.drop(columns='label', axis=1)
    df_y_2018 = df_2018[['label']]
    df_X_2019 = df_2019.drop(columns='label', axis=1)
    df_y_2019 = df_2019[['label']]
    

    ## UnderSampler ##

    from imblearn.under_sampling import RandomUnderSampler

    X_samp_11, y_samp_11 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2011, df_y_2011)
    X_samp_12, y_samp_12 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2012, df_y_2012)
    X_samp_13, y_samp_13 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2013, df_y_2013)
    X_samp_14, y_samp_14 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2014, df_y_2014)
    X_samp_15, y_samp_15 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2015, df_y_2015)
    X_samp_16, y_samp_16 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2016, df_y_2016)
    X_samp_17, y_samp_17 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2017, df_y_2017)
    X_samp_18, y_samp_18 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2018, df_y_2018)
    X_samp_19, y_samp_19 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2019, df_y_2019)

    X_samp = pd.concat([X_samp_11, X_samp_12, X_samp_13, X_samp_14, X_samp_15, X_samp_16, X_samp_17, X_samp_18, X_samp_19], axis=0)
    y_samp = pd.concat([y_samp_11, y_samp_12, y_samp_13, y_samp_14, y_samp_15, y_samp_16, y_samp_17, y_samp_18, y_samp_19], axis=0)

    a = pd.concat([X_samp, y_samp], axis=1)

    # 교차검증 3 데이터 셋
    X_samp_train = a[(a['회계년도'] <= '2019/12') & (a['회계년도'] >= '2017/12') | (a['회계년도'] >= '2011/12') & (a['회계년도'] <= '2014/12')]
    X_samp_test = a[(a['회계년도'] <= '2016/12') & (a['회계년도'] >= '2015/12')]

    X_samp_train_1 = X_samp_train.drop(columns=['label', '회계년도'], axis=1)
    y_samp_train_1 = X_samp_train[['label']]

    X_samp_test_1 = X_samp_test.drop(columns=['label', '회계년도'], axis=1)
    y_samp_test_1 = X_samp_test[['label']]

    X_samp_train_1.reset_index(drop=True, inplace=True)
    X_samp_test_1.reset_index(drop=True, inplace=True)
    y_samp_train_1.reset_index(drop=True, inplace=True)
    y_samp_test_1.reset_index(drop=True, inplace=True)

    from imblearn.over_sampling import BorderlineSMOTE

    X_samp_2, y_samp_2 = BorderlineSMOTE(random_state=0,sampling_strategy=1).fit_resample(X_samp_train_1, y_samp_train_1)

    train_df = pd.concat([X_samp_2, y_samp_2], axis=1)
    test_df = pd.concat([X_samp_test_1, y_samp_test_1], axis=1)

    return train_df, test_df

model5(kk)[0]['label'].value_counts()

model5(kk)[1]['label'].value_counts()

# 6번 데이터 검증

def model6(Data):

    train = Data[(Data['회계년도'] <= '2019/12') & (Data['회계년도'] >= '2018/12') | (Data['회계년도'] >= '2011/12') & (Data['회계년도'] <= '2015/12')].reset_index(drop=True)
    test = Data[(Data['회계년도'] <= '2017/12') & (Data['회계년도'] >= '2016/12')].reset_index(drop=True)

    X_train = train.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_train_concat = train[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_train = train[['label']]
    X_test = test.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_test_concat = test[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_test = test[['label']]


    ## Scaling##

    from sklearn.preprocessing import StandardScaler
    # StandardScaler() Scaler객체 생성.
    scaler = StandardScaler()
    # 학습 데이터에 대해서 fit(), transform() 수행.
    scaler.fit(X_train)
    train_scaled_SS = scaler.transform(X_train)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_train_SS = pd.DataFrame(data = train_scaled_SS, columns=X_train.columns)

    # 테스트 데이터에서는 다시 fit(), transform()이나 fit_transform()을 수행하지 않고 transform만 수행.
    test_scaled_MM = scaler.transform(X_test)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_test_SS = pd.DataFrame(data = test_scaled_MM, columns=X_test.columns)


    X_train_SS = pd.concat([X_train_SS,X_train_concat,y_train], axis=1)
    X_test_SS = pd.concat([X_test_SS,X_test_concat, y_test], axis=1)

    trian_test_df = pd.concat([X_train_SS, X_test_SS], axis=0)
    trian_test_df.reset_index(drop=True, inplace=True)      

    df_2011 = trian_test_df[trian_test_df['회계년도'] == '2011/12']
    df_2012 = trian_test_df[trian_test_df['회계년도'] == '2012/12']
    df_2013 = trian_test_df[trian_test_df['회계년도'] == '2013/12']
    df_2014 = trian_test_df[trian_test_df['회계년도'] == '2014/12']
    df_2015 = trian_test_df[trian_test_df['회계년도'] == '2015/12']
    df_2016 = trian_test_df[trian_test_df['회계년도'] == '2016/12']
    df_2017 = trian_test_df[trian_test_df['회계년도'] == '2017/12']
    df_2018 = trian_test_df[trian_test_df['회계년도'] == '2018/12']
    df_2019 = trian_test_df[trian_test_df['회계년도'] == '2019/12']

    df_X_2011 = df_2011.drop(columns='label', axis=1)
    df_y_2011 = df_2011[['label']]
    df_X_2012 = df_2012.drop(columns='label', axis=1)
    df_y_2012 = df_2012[['label']]
    df_X_2013 = df_2013.drop(columns='label', axis=1)
    df_y_2013 = df_2013[['label']]
    df_X_2014 = df_2014.drop(columns='label', axis=1)
    df_y_2014 = df_2014[['label']]
    df_X_2015 = df_2015.drop(columns='label', axis=1)
    df_y_2015 = df_2015[['label']]
    df_X_2016 = df_2016.drop(columns='label', axis=1)
    df_y_2016 = df_2016[['label']]
    df_X_2017 = df_2017.drop(columns='label', axis=1)
    df_y_2017 = df_2017[['label']]
    df_X_2018 = df_2018.drop(columns='label', axis=1)
    df_y_2018 = df_2018[['label']]
    df_X_2019 = df_2019.drop(columns='label', axis=1)
    df_y_2019 = df_2019[['label']]
    

    ## UnderSampler ##

    from imblearn.under_sampling import RandomUnderSampler

    X_samp_11, y_samp_11 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2011, df_y_2011)
    X_samp_12, y_samp_12 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2012, df_y_2012)
    X_samp_13, y_samp_13 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2013, df_y_2013)
    X_samp_14, y_samp_14 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2014, df_y_2014)
    X_samp_15, y_samp_15 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2015, df_y_2015)
    X_samp_16, y_samp_16 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2016, df_y_2016)
    X_samp_17, y_samp_17 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2017, df_y_2017)
    X_samp_18, y_samp_18 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2018, df_y_2018)
    X_samp_19, y_samp_19 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2019, df_y_2019)

    X_samp = pd.concat([X_samp_11, X_samp_12, X_samp_13, X_samp_14, X_samp_15, X_samp_16, X_samp_17, X_samp_18, X_samp_19], axis=0)
    y_samp = pd.concat([y_samp_11, y_samp_12, y_samp_13, y_samp_14, y_samp_15, y_samp_16, y_samp_17, y_samp_18, y_samp_19], axis=0)

    a = pd.concat([X_samp, y_samp], axis=1)

    # 교차검증 3 데이터 셋
    X_samp_train = a[(a['회계년도'] <= '2019/12') & (a['회계년도'] >= '2018/12') | (a['회계년도'] >= '2011/12') & (a['회계년도'] <= '2015/12')]
    X_samp_test = a[(a['회계년도'] <= '2017/12') & (a['회계년도'] >= '2016/12')]

    X_samp_train_1 = X_samp_train.drop(columns=['label', '회계년도'], axis=1)
    y_samp_train_1 = X_samp_train[['label']]

    X_samp_test_1 = X_samp_test.drop(columns=['label', '회계년도'], axis=1)
    y_samp_test_1 = X_samp_test[['label']]

    X_samp_train_1.reset_index(drop=True, inplace=True)
    X_samp_test_1.reset_index(drop=True, inplace=True)
    y_samp_train_1.reset_index(drop=True, inplace=True)
    y_samp_test_1.reset_index(drop=True, inplace=True)

    from imblearn.over_sampling import BorderlineSMOTE

    X_samp_2, y_samp_2 = BorderlineSMOTE(random_state=0,sampling_strategy=1).fit_resample(X_samp_train_1, y_samp_train_1)

    train_df = pd.concat([X_samp_2, y_samp_2], axis=1)
    test_df = pd.concat([X_samp_test_1, y_samp_test_1], axis=1)

    return train_df, test_df

model6(kk)[0]['label'].value_counts()

model6(kk)[1]['label'].value_counts()

# 7번 데이터 검증

def model7(Data):

    train = Data[(Data['회계년도'] == '2019/12') | (Data['회계년도'] >= '2011/12') & (Data['회계년도'] <= '2016/12')].reset_index(drop=True)
    test = Data[(Data['회계년도'] <= '2018/12') & (Data['회계년도'] >= '2017/12')].reset_index(drop=True)

    X_train = train.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_train_concat = train[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_train = train[['label']]
    X_test = test.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_test_concat = test[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_test = test[['label']]


    ## Scaling##

    from sklearn.preprocessing import StandardScaler
    # StandardScaler() Scaler객체 생성.
    scaler = StandardScaler()
    # 학습 데이터에 대해서 fit(), transform() 수행.
    scaler.fit(X_train)
    train_scaled_SS = scaler.transform(X_train)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_train_SS = pd.DataFrame(data = train_scaled_SS, columns=X_train.columns)

    # 테스트 데이터에서는 다시 fit(), transform()이나 fit_transform()을 수행하지 않고 transform만 수행.
    test_scaled_MM = scaler.transform(X_test)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_test_SS = pd.DataFrame(data = test_scaled_MM, columns=X_test.columns)


    X_train_SS = pd.concat([X_train_SS,X_train_concat,y_train], axis=1)
    X_test_SS = pd.concat([X_test_SS,X_test_concat, y_test], axis=1)

    trian_test_df = pd.concat([X_train_SS, X_test_SS], axis=0)
    trian_test_df.reset_index(drop=True, inplace=True)      

    df_2011 = trian_test_df[trian_test_df['회계년도'] == '2011/12']
    df_2012 = trian_test_df[trian_test_df['회계년도'] == '2012/12']
    df_2013 = trian_test_df[trian_test_df['회계년도'] == '2013/12']
    df_2014 = trian_test_df[trian_test_df['회계년도'] == '2014/12']
    df_2015 = trian_test_df[trian_test_df['회계년도'] == '2015/12']
    df_2016 = trian_test_df[trian_test_df['회계년도'] == '2016/12']
    df_2017 = trian_test_df[trian_test_df['회계년도'] == '2017/12']
    df_2018 = trian_test_df[trian_test_df['회계년도'] == '2018/12']
    df_2019 = trian_test_df[trian_test_df['회계년도'] == '2019/12']

    df_X_2011 = df_2011.drop(columns='label', axis=1)
    df_y_2011 = df_2011[['label']]
    df_X_2012 = df_2012.drop(columns='label', axis=1)
    df_y_2012 = df_2012[['label']]
    df_X_2013 = df_2013.drop(columns='label', axis=1)
    df_y_2013 = df_2013[['label']]
    df_X_2014 = df_2014.drop(columns='label', axis=1)
    df_y_2014 = df_2014[['label']]
    df_X_2015 = df_2015.drop(columns='label', axis=1)
    df_y_2015 = df_2015[['label']]
    df_X_2016 = df_2016.drop(columns='label', axis=1)
    df_y_2016 = df_2016[['label']]
    df_X_2017 = df_2017.drop(columns='label', axis=1)
    df_y_2017 = df_2017[['label']]
    df_X_2018 = df_2018.drop(columns='label', axis=1)
    df_y_2018 = df_2018[['label']]
    df_X_2019 = df_2019.drop(columns='label', axis=1)
    df_y_2019 = df_2019[['label']]
    

    ## UnderSampler ##

    from imblearn.under_sampling import RandomUnderSampler

    X_samp_11, y_samp_11 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2011, df_y_2011)
    X_samp_12, y_samp_12 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2012, df_y_2012)
    X_samp_13, y_samp_13 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2013, df_y_2013)
    X_samp_14, y_samp_14 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2014, df_y_2014)
    X_samp_15, y_samp_15 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2015, df_y_2015)
    X_samp_16, y_samp_16 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2016, df_y_2016)
    X_samp_17, y_samp_17 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2017, df_y_2017)
    X_samp_18, y_samp_18 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2018, df_y_2018)
    X_samp_19, y_samp_19 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2019, df_y_2019)

    X_samp = pd.concat([X_samp_11, X_samp_12, X_samp_13, X_samp_14, X_samp_15, X_samp_16, X_samp_17, X_samp_18, X_samp_19], axis=0)
    y_samp = pd.concat([y_samp_11, y_samp_12, y_samp_13, y_samp_14, y_samp_15, y_samp_16, y_samp_17, y_samp_18, y_samp_19], axis=0)

    a = pd.concat([X_samp, y_samp], axis=1)

    # 교차검증 3 데이터 셋
    X_samp_train = a[(a['회계년도'] == '2019/12') | (a['회계년도'] >= '2011/12') & (a['회계년도'] <= '2016/12')]
    X_samp_test = a[(a['회계년도'] <= '2018/12') & (a['회계년도'] >= '2017/12')]

    X_samp_train_1 = X_samp_train.drop(columns=['label', '회계년도'], axis=1)
    y_samp_train_1 = X_samp_train[['label']]

    X_samp_test_1 = X_samp_test.drop(columns=['label', '회계년도'], axis=1)
    y_samp_test_1 = X_samp_test[['label']]

    X_samp_train_1.reset_index(drop=True, inplace=True)
    X_samp_test_1.reset_index(drop=True, inplace=True)
    y_samp_train_1.reset_index(drop=True, inplace=True)
    y_samp_test_1.reset_index(drop=True, inplace=True)

    from imblearn.over_sampling import BorderlineSMOTE

    X_samp_2, y_samp_2 = BorderlineSMOTE(random_state=0,sampling_strategy=1).fit_resample(X_samp_train_1, y_samp_train_1)

    train_df = pd.concat([X_samp_2, y_samp_2], axis=1)
    test_df = pd.concat([X_samp_test_1, y_samp_test_1], axis=1)

    return train_df, test_df

model7(kk)[0]['label'].value_counts()

model7(kk)[1]['label'].value_counts()

# 0번 데이터 검증

def model0(Data):

    train = Data[(Data['회계년도'] <= '2017/12') & (Data['회계년도'] >= '2011/12')].reset_index(drop=True)
    test = Data[(Data['회계년도'] <= '2019/12') & (Data['회계년도'] >= '2018/12')].reset_index(drop=True)

    ## Train Scaling 전 Data 분리하기 ##

    X_train = train.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_train_concat = train[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_train = train[['label']]
    X_test = test.drop(columns=['label', '산업분류', '기업생애주기_Shake-Out',
       '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도'], axis=1)
    X_test_concat = test[['산업분류', '기업생애주기_Shake-Out',
        '기업생애주기_도입기', '기업생애주기_성숙기', '기업생애주기_성장기', '기업생애주기_쇠퇴기', '회계년도']]
    y_test = test[['label']]


    ## Scaling##

    from sklearn.preprocessing import StandardScaler
    # StandardScaler() Scaler객체 생성.
    scaler = StandardScaler()
    # 학습 데이터에 대해서 fit(), transform() 수행.
    scaler.fit(X_train)
    train_scaled_SS = scaler.transform(X_train)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_train_SS = pd.DataFrame(data = train_scaled_SS, columns=X_train.columns)

    # 테스트 데이터에서는 다시 fit(), transform()이나 fit_transform()을 수행하지 않고 transform만 수행.
    test_scaled_MM = scaler.transform(X_test)
    # transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
    X_test_SS = pd.DataFrame(data = test_scaled_MM, columns=X_test.columns)


    X_train_SS = pd.concat([X_train_SS,X_train_concat,y_train], axis=1)
    X_test_SS = pd.concat([X_test_SS,X_test_concat, y_test], axis=1)

    trian_test_df = pd.concat([X_train_SS, X_test_SS], axis=0)
    trian_test_df.reset_index(drop=True, inplace=True)      

    df_2011 = trian_test_df[trian_test_df['회계년도'] == '2011/12']
    df_2012 = trian_test_df[trian_test_df['회계년도'] == '2012/12']
    df_2013 = trian_test_df[trian_test_df['회계년도'] == '2013/12']
    df_2014 = trian_test_df[trian_test_df['회계년도'] == '2014/12']
    df_2015 = trian_test_df[trian_test_df['회계년도'] == '2015/12']
    df_2016 = trian_test_df[trian_test_df['회계년도'] == '2016/12']
    df_2017 = trian_test_df[trian_test_df['회계년도'] == '2017/12']
    df_2018 = trian_test_df[trian_test_df['회계년도'] == '2018/12']
    df_2019 = trian_test_df[trian_test_df['회계년도'] == '2019/12']

    df_X_2011 = df_2011.drop(columns='label', axis=1)
    df_y_2011 = df_2011[['label']]
    df_X_2012 = df_2012.drop(columns='label', axis=1)
    df_y_2012 = df_2012[['label']]
    df_X_2013 = df_2013.drop(columns='label', axis=1)
    df_y_2013 = df_2013[['label']]
    df_X_2014 = df_2014.drop(columns='label', axis=1)
    df_y_2014 = df_2014[['label']]
    df_X_2015 = df_2015.drop(columns='label', axis=1)
    df_y_2015 = df_2015[['label']]
    df_X_2016 = df_2016.drop(columns='label', axis=1)
    df_y_2016 = df_2016[['label']]
    df_X_2017 = df_2017.drop(columns='label', axis=1)
    df_y_2017 = df_2017[['label']]
    df_X_2018 = df_2018.drop(columns='label', axis=1)
    df_y_2018 = df_2018[['label']]
    df_X_2019 = df_2019.drop(columns='label', axis=1)
    df_y_2019 = df_2019[['label']]
    

    ## UnderSampler ##

    from imblearn.under_sampling import RandomUnderSampler

    X_samp_11, y_samp_11 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2011, df_y_2011)
    X_samp_12, y_samp_12 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2012, df_y_2012)
    X_samp_13, y_samp_13 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2013, df_y_2013)
    X_samp_14, y_samp_14 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2014, df_y_2014)
    X_samp_15, y_samp_15 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2015, df_y_2015)
    X_samp_16, y_samp_16 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2016, df_y_2016)
    X_samp_17, y_samp_17 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2017, df_y_2017)
    X_samp_18, y_samp_18 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2018, df_y_2018)
    X_samp_19, y_samp_19 = RandomUnderSampler(random_state=0,sampling_strategy=0.2).fit_resample(df_X_2019, df_y_2019)

    X_samp = pd.concat([X_samp_11, X_samp_12, X_samp_13, X_samp_14, X_samp_15, X_samp_16, X_samp_17, X_samp_18, X_samp_19], axis=0)
    y_samp = pd.concat([y_samp_11, y_samp_12, y_samp_13, y_samp_14, y_samp_15, y_samp_16, y_samp_17, y_samp_18, y_samp_19], axis=0)

    trian_test_samp_df = pd.concat([X_samp, y_samp], axis=1)

    # 교차검증 1 데이터 셋
    X_samp_train = trian_test_samp_df[(trian_test_samp_df['회계년도'] <= '2017/12') & (trian_test_samp_df['회계년도'] >= '2011/12')]
    X_samp_test = trian_test_samp_df[(trian_test_samp_df['회계년도'] <= '2019/12') & (trian_test_samp_df['회계년도'] >= '2018/12')]

    X_samp_train_1 = X_samp_train.drop(columns=['label', '회계년도'], axis=1)
    y_samp_train_1 = X_samp_train[['label']]

    X_samp_test_1 = X_samp_test.drop(columns=['label', '회계년도'], axis=1)
    y_samp_test_1 = X_samp_test[['label']]

    X_samp_train_1.reset_index(drop=True, inplace=True)
    X_samp_test_1.reset_index(drop=True, inplace=True)
    y_samp_train_1.reset_index(drop=True, inplace=True)
    y_samp_test_1.reset_index(drop=True, inplace=True)

    from imblearn.over_sampling import BorderlineSMOTE

    X_samp_2, y_samp_2 = BorderlineSMOTE(random_state=0,sampling_strategy=1).fit_resample(X_samp_train_1, y_samp_train_1)

    train_df = pd.concat([X_samp_2, y_samp_2], axis=1)
    test_df = pd.concat([X_samp_test_1, y_samp_test_1], axis=1)

    return train_df, test_df

model0(kk)[0]['label'].value_counts()

model0(kk)[1]['label'].value_counts()

X_train_1 = model1(kk)[0].drop(columns=['label'], axis=1)
y_train_1 = model1(kk)[0][['label']]
X_test_1 = model1(kk)[1].drop(columns=['label'], axis=1)
y_test_1 = model1(kk)[1][['label']]

X_train_2 = model2(kk)[0].drop(columns=['label'], axis=1)
y_train_2 = model2(kk)[0][['label']]
X_test_2 = model2(kk)[1].drop(columns=['label'], axis=1)
y_test_2 = model2(kk)[1][['label']]

X_train_3 = model3(kk)[0].drop(columns=['label'], axis=1)
y_train_3 = model3(kk)[0][['label']]
X_test_3 = model3(kk)[1].drop(columns=['label'], axis=1)
y_test_3 = model3(kk)[1][['label']]

X_train_4 = model4(kk)[0].drop(columns=['label'], axis=1)
y_train_4 = model4(kk)[0][['label']]
X_test_4 = model4(kk)[1].drop(columns=['label'], axis=1)
y_test_4 = model4(kk)[1][['label']]

X_train_5 = model5(kk)[0].drop(columns=['label'], axis=1)
y_train_5 = model5(kk)[0][['label']]
X_test_5 = model5(kk)[1].drop(columns=['label'], axis=1)
y_test_5 = model5(kk)[1][['label']]

X_train_6 = model6(kk)[0].drop(columns=['label'], axis=1)
y_train_6 = model6(kk)[0][['label']]
X_test_6 = model6(kk)[1].drop(columns=['label'], axis=1)
y_test_6 = model6(kk)[1][['label']]

X_train_7 = model7(kk)[0].drop(columns=['label'], axis=1)
y_train_7 = model7(kk)[0][['label']]
X_test_7 = model7(kk)[1].drop(columns=['label'], axis=1)
y_test_7 = model7(kk)[1][['label']]

def model_basic(x_train, y_train, x_test, y_test): 
    models = [
        LogisticRegression(random_state=0),
        SVC(random_state=0),
        DecisionTreeClassifier(random_state=0),
        RandomForestClassifier(random_state=0),
        XGBClassifier(random_state=0),
        LGBMClassifier(random_state=0) 
    ]

    rdict={'model':[],'acc_train':[], 'auc_train':[], 'acc_test':[],'precision':[],'recall':[],'f1_score':[], 'AUC_test':[]}


    for clf in models:
        clf = clf.fit(x_train, y_train)
    #1열:Train
        y_hat = clf.predict(x_train)
        results_train  = (round(accuracy_score(y_train,y_hat),2),round(roc_auc_score(y_train,y_hat),2))
    #2열:Test
        y_hat = clf.predict(x_test)
        results = (round(accuracy_score(y_test,y_hat),2),
                        round(precision_score(y_test,y_hat),2),
                        round(recall_score(y_test,y_hat),2),
                        round(f1_score(y_test,y_hat),2),
                        round(roc_auc_score(y_test,y_hat),2))

        rdict['model'].append(clf); 
        rdict['acc_train'].append(results_train[0])
        rdict['auc_train'].append(results_train[1])
        
        rdict['acc_test'].append(results[0])
        rdict['precision'].append(results[1])
        rdict['recall'].append(results[2])
        rdict['f1_score'].append(results[3])
        rdict['AUC_test'].append(results[4])   

        # confusion = confusion_matrix(y_test, y_hat)

        # print(confusion)

    rdf_final = pd.DataFrame(data=rdict)
    return rdf_final

a_0 = model_basic(X_train_1, y_train_1, X_test_1, y_test_1).iloc[0]
a_1 = model_basic(X_train_1, y_train_1, X_test_1, y_test_1).iloc[1]
a_2 = model_basic(X_train_1, y_train_1, X_test_1, y_test_1).iloc[2]
a_3 = model_basic(X_train_1, y_train_1, X_test_1, y_test_1).iloc[3]
a_4 = model_basic(X_train_1, y_train_1, X_test_1, y_test_1).iloc[4]
a_5 = model_basic(X_train_1, y_train_1, X_test_1, y_test_1).iloc[5]

b_0 = model_basic(X_train_2, y_train_2, X_test_2, y_test_2).iloc[0]
b_1 = model_basic(X_train_2, y_train_2, X_test_2, y_test_2).iloc[1]
b_2 = model_basic(X_train_2, y_train_2, X_test_2, y_test_2).iloc[2]
b_3 = model_basic(X_train_2, y_train_2, X_test_2, y_test_2).iloc[3]
b_4 = model_basic(X_train_2, y_train_2, X_test_2, y_test_2).iloc[4]
b_5 = model_basic(X_train_2, y_train_2, X_test_2, y_test_2).iloc[5]

c_0 = model_basic(X_train_3, y_train_3, X_test_3, y_test_3).iloc[0]
c_1 = model_basic(X_train_3, y_train_3, X_test_3, y_test_3).iloc[1]
c_2 = model_basic(X_train_3, y_train_3, X_test_3, y_test_3).iloc[2]
c_3 = model_basic(X_train_3, y_train_3, X_test_3, y_test_3).iloc[3]
c_4 = model_basic(X_train_3, y_train_3, X_test_3, y_test_3).iloc[4]
c_5 = model_basic(X_train_3, y_train_3, X_test_3, y_test_3).iloc[5]

d_0 = model_basic(X_train_4, y_train_4, X_test_4, y_test_4).iloc[0]
d_1 = model_basic(X_train_4, y_train_4, X_test_4, y_test_4).iloc[1]
d_2 = model_basic(X_train_4, y_train_4, X_test_4, y_test_4).iloc[2]
d_3 = model_basic(X_train_4, y_train_4, X_test_4, y_test_4).iloc[3]
d_4 = model_basic(X_train_4, y_train_4, X_test_4, y_test_4).iloc[4]
d_5 = model_basic(X_train_4, y_train_4, X_test_4, y_test_4).iloc[5]

e_0 = model_basic(X_train_5, y_train_5, X_test_5, y_test_5).iloc[0]
e_1 = model_basic(X_train_5, y_train_5, X_test_5, y_test_5).iloc[1]
e_2 = model_basic(X_train_5, y_train_5, X_test_5, y_test_5).iloc[2]
e_3 = model_basic(X_train_5, y_train_5, X_test_5, y_test_5).iloc[3]
e_4 = model_basic(X_train_5, y_train_5, X_test_5, y_test_5).iloc[4]
e_5 = model_basic(X_train_5, y_train_5, X_test_5, y_test_5).iloc[5]

f_0 = model_basic(X_train_6, y_train_6, X_test_6, y_test_6).iloc[0]
f_1 = model_basic(X_train_6, y_train_6, X_test_6, y_test_6).iloc[1]
f_2 = model_basic(X_train_6, y_train_6, X_test_6, y_test_6).iloc[2]
f_3 = model_basic(X_train_6, y_train_6, X_test_6, y_test_6).iloc[3]
f_4 = model_basic(X_train_6, y_train_6, X_test_6, y_test_6).iloc[4]
f_5 = model_basic(X_train_6, y_train_6, X_test_6, y_test_6).iloc[5]

g_0 = model_basic(X_train_7, y_train_7, X_test_7, y_test_7).iloc[0]
g_1 = model_basic(X_train_7, y_train_7, X_test_7, y_test_7).iloc[1]
g_2 = model_basic(X_train_7, y_train_7, X_test_7, y_test_7).iloc[2]
g_3 = model_basic(X_train_7, y_train_7, X_test_7, y_test_7).iloc[3]
g_4 = model_basic(X_train_7, y_train_7, X_test_7, y_test_7).iloc[4]
g_5 = model_basic(X_train_7, y_train_7, X_test_7, y_test_7).iloc[5]

df_0 = pd.DataFrame(a_0)
df_1 = pd.DataFrame(b_0)
df_2 = pd.DataFrame(c_0)
df_3 = pd.DataFrame(d_0)
df_4 = pd.DataFrame(e_0)
df_5 = pd.DataFrame(f_0)
df_6 = pd.DataFrame(e_0)

df = pd.concat([df_0, df_1, df_2, df_3, df_4, df_5, df_6], axis=1)
# 첫 번째 행을 컬럼으로 옮기기
df.columns = df.iloc[0]
df = df.iloc[1:]
dfdf1 = pd.DataFrame(df.T.mean())
dfdf1.columns = ['Logistic']
dfdf1

df_0 = pd.DataFrame(a_1)
df_1 = pd.DataFrame(b_1)
df_2 = pd.DataFrame(c_1)
df_3 = pd.DataFrame(d_1)
df_4 = pd.DataFrame(e_1)
df_5 = pd.DataFrame(f_1)
df_6 = pd.DataFrame(e_1)

df = pd.concat([df_0, df_1, df_2, df_3, df_4, df_5, df_6], axis=1)
# 첫 번째 행을 컬럼으로 옮기기
df.columns = df.iloc[0]
df = df.iloc[1:]
dfdf2 = pd.DataFrame(df.T.mean())
dfdf2.columns = ['SVC']
dfdf2

df_0 = pd.DataFrame(a_2)
df_1 = pd.DataFrame(b_2)
df_2 = pd.DataFrame(c_2)
df_3 = pd.DataFrame(d_2)
df_4 = pd.DataFrame(e_2)
df_5 = pd.DataFrame(f_2)
df_6 = pd.DataFrame(e_2)

df = pd.concat([df_0, df_1, df_2, df_3, df_4, df_5, df_6], axis=1)
# 첫 번째 행을 컬럼으로 옮기기
df.columns = df.iloc[0]
df = df.iloc[1:]
dfdf3 = pd.DataFrame(df.T.mean())
dfdf3.columns = ['DecisionTree']
dfdf3

df_0 = pd.DataFrame(a_3)
df_1 = pd.DataFrame(b_3)
df_2 = pd.DataFrame(c_3)
df_3 = pd.DataFrame(d_3)
df_4 = pd.DataFrame(e_3)
df_5 = pd.DataFrame(f_3)
df_6 = pd.DataFrame(e_3)

df = pd.concat([df_0, df_1, df_2, df_3, df_4, df_5, df_6], axis=1)
# 첫 번째 행을 컬럼으로 옮기기
df.columns = df.iloc[0]
df = df.iloc[1:]
dfdf4 = pd.DataFrame(df.T.mean())
dfdf4.columns = ['RandomForest']
dfdf4

df_0 = pd.DataFrame(a_4)
df_1 = pd.DataFrame(b_4)
df_2 = pd.DataFrame(c_4)
df_3 = pd.DataFrame(d_4)
df_4 = pd.DataFrame(e_4)
df_5 = pd.DataFrame(f_4)
df_6 = pd.DataFrame(e_4)

df = pd.concat([df_0, df_1, df_2, df_3, df_4, df_5, df_6], axis=1)
# 첫 번째 행을 컬럼으로 옮기기
df.columns = df.iloc[0]
df = df.iloc[1:]
dfdf5 = pd.DataFrame(df.T.mean())
dfdf5.columns = ['XGB']
dfdf5

df_0 = pd.DataFrame(a_5)
df_1 = pd.DataFrame(b_5)
df_2 = pd.DataFrame(c_5)
df_3 = pd.DataFrame(d_5)
df_4 = pd.DataFrame(e_5)
df_5 = pd.DataFrame(f_5)
df_6 = pd.DataFrame(e_5)

df = pd.concat([df_0, df_1, df_2, df_3, df_4, df_5, df_6], axis=1)
# 첫 번째 행을 컬럼으로 옮기기
df.columns = df.iloc[0]
df = df.iloc[1:]
dfdf6 = pd.DataFrame(df.T.mean())
dfdf6.columns = ['LGBM']
dfdf6

kosdaq_교차검증평균 = round(pd.concat([dfdf1,dfdf2,dfdf3,dfdf4,dfdf5,dfdf6], axis=1),4)
kosdaq_교차검증평균.T
