==> 9. DS2(코스피+코스닥)_투자Filter.ipynb <==
import pandas as pd
import numpy as np

# 필터링 기준 만들기
1. 교수님 조언 : 3년 연속 순이익이 증가한 기업
2. 시가총액 하위 20% 거르기

# 1. 3년 연속 순이익이 증가한 기업

### 1-1 코스피

# 데이터 로드
ni_kospi = pd.read_csv("./datasets/kospi_영업이익관련(종진형님).csv", encoding='cp949')
ni_kosdaq = pd.read_csv("./datasets/kosdaq_영업이익관련(종진형님).csv", encoding='cp949')

# 코스피 컬럼명 변경
ni_kospi.columns = ['회사명', '거래소코드', '회계년도', '매출총이익(연결)', '당기순이익(연결)', '영업이익(연결)',
       '매출총이익(개별)', '당기순이익(개별)','영업이익(개별)']

# 결측치 확인
ni_kospi.isna().sum()

# 연결
ni_kospi_ifrs = ni_kospi[['매출총이익(연결)', '당기순이익(연결)', '영업이익(연결)']]
ni_kospi_ifrs.columns = [['매출총이익', '당기순이익', '영업이익']]
ni_kospi_ifrs

# 개별
ni_kospi_gaap = ni_kospi[['매출총이익(개별)', '당기순이익(개별)','영업이익(개별)']]
ni_kospi_gaap.columns = [['매출총이익', '당기순이익', '영업이익']]
ni_kospi_gaap

# 회사명 거래소코드 회계년도 따로 저장
ni_kospi_info = ni_kospi[['회사명', '거래소코드', '회계년도']]

# 연결 결측치 개별로 채우는 코드
ni_kospi_combined = ni_kospi_ifrs.combine_first(ni_kospi_gaap)
ni_kospi_combined.isna().sum()

kospi_revenue = pd.concat([ni_kospi_info, ni_kospi_combined], axis=1)
kospi_revenue.columns = ['회사명','거래소코드','회계년도','매출총이익','당기순이익','영업이익']
kospi_revenue

### 1-2 코스닥

# 컬럼명 변경
ni_kosdaq.columns = ['회사명', '거래소코드', '회계년도', '매출총이익(연결)', '당기순이익(연결)', '영업이익(연결)',
       '매출총이익(개별)', '당기순이익(개별)','영업이익(개별)']

ni_kosdaq.isna().sum()

ni_kosdaq_ifrs = ni_kosdaq[['매출총이익(연결)', '당기순이익(연결)', '영업이익(연결)']]
ni_kosdaq_ifrs.columns = [['매출총이익', '당기순이익', '영업이익']]
ni_kosdaq_ifrs

ni_kosdaq_gaap = ni_kosdaq[['매출총이익(개별)', '당기순이익(개별)','영업이익(개별)']]
ni_kosdaq_gaap.columns = [['매출총이익', '당기순이익', '영업이익']]
ni_kosdaq_gaap

ni_kosdaq_info = ni_kosdaq[['회사명', '거래소코드', '회계년도']]

ni_kosdaq_combined = ni_kosdaq_ifrs.combine_first(ni_kosdaq_gaap)
ni_kosdaq_combined

kosdaq_revenue = pd.concat([ni_kosdaq_info, ni_kosdaq_combined], axis=1)
kosdaq_revenue.columns = ['회사명','거래소코드','회계년도','매출총이익','당기순이익','영업이익']

revenue = pd.concat([kospi_revenue, kosdaq_revenue], axis=0)

revenue.drop_duplicates(inplace=True)

revenue.drop(labels='매출총이익', axis=1, inplace=True)

# 결측치 제거
revenue.dropna(inplace=True)
revenue

revenue['3년당기순이익흑자'] = np.where((revenue['당기순이익'] > 0) &
                            (revenue['당기순이익'].shift(1) > 0) &
                            (revenue['당기순이익'].shift(2) > 0), 1, 0)

revenue['3년영업이익흑자'] = np.where((revenue['영업이익'] > 0) &
                            (revenue['영업이익'].shift(1) > 0) &
                            (revenue['영업이익'].shift(2) > 0), 1, 0)

revenue[revenue['3년당기순이익흑자'] == 1]

revenue.reset_index(drop=True, inplace=True)

revenue

# 2. 이상치 처리전 데이터 불러오기

kosdaq = pd.read_csv('./datasets/2.kosdaq_결측치제거완료.csv' ,index_col=0)
kospi = pd.read_csv('./datasets/2.kospi_결측치제거완료.csv', index_col=0)

kosdaqpi = pd.concat([kospi, kosdaq], axis=0)

kosdaqpi.drop_duplicates(subset=['거래소코드', '회계년도'], keep='first', inplace=True)

kosdaqpi[kosdaqpi['회사명'] == '(주)더블유게임즈']

df = pd.merge(revenue, kosdaqpi, how = 'left', on=['회사명','거래소코드', '회계년도'])
df

# 결과창 늘려주는 코드 시작
pd.set_option('display.max_rows',None)
df.isna().sum()

# 결과창 늘려주는 코드 종료
pd.reset_option('display.max_rows', None)

df = df[~df['총자본증가율'].isna()]
df.reset_index(drop=True, inplace=True)

df.info(verbose=True)

filtered = df[df['3년영업이익흑자'] == 1]
filtered

filtered = filtered[['회사명','거래소코드','회계년도','총자본증가율','총자본순이익률','자기자본순이익률','매출원가대매출액비율','금융비용대총비용비율',
 '자기자본구성비율','순운전자본비율','CASHFLOW대총자본비율','순운전자본회전률']]
filtered

filtered.reset_index(drop=True, inplace=True)

filtered

# 3. 시가총액 붙이기

시가총액 = pd.read_csv('./datasets/kospi+kosdaq_시가총액.csv', index_col=0)
시가총액

# 종목코드를 6자리로 통일. ex) 990 -> 000990
def padding(x):
    x = str(x)
    return x.rjust(6, "0")

filtered['거래소코드'] = filtered['거래소코드'].map(padding)

merge_df = pd.merge(filtered, 시가총액, how='left', on=['거래소코드', '회계년도'])
merge_df

merge_df.drop(columns=['종가','상장주식수'], axis=1, inplace=True)
merge_df

merge_df.to_csv('./datasets/투자/KOSDAQ+KOSPI_DATA(투자).csv')

# 3년연속 흑자기업 선별 + 시가총액 컬럼 생성 완료
