==> 12. DS1,DS2_Feature비교.ipynb <==
import pandas as pd

# KOSPI+KOSDAQ

train_df = pd.read_csv('./datasets/KOSDAQ+KOSPI_TRAIN_DATA.csv', index_col=0)
train_df = train_df.iloc[:,2:]
test_df = pd.read_csv('./datasets/KOSDAQ+KOSPI_TEST1_DATA.csv', index_col=0)
test_df = test_df.iloc[:,2:]

train_df = train_df[['총자본증가율',
 '총자본순이익률',
 '자기자본순이익률',
 '매출원가대매출액비율',
 '금융비용대총비용비율',
 '자기자본구성비율',
 '순운전자본비율',
 'CASHFLOW대총자본비율',
 '순운전자본회전률', 'label', '1주당순이익', '1주당정상영업이익']]

round(train_df[train_df['label'] == 1].describe().T,4)

round(train_df[train_df['label'] == 0].describe().T,4)

test_df = test_df[['총자본증가율',
 '총자본순이익률',
 '자기자본순이익률',
 '매출원가대매출액비율',
 '금융비용대총비용비율',
 '자기자본구성비율',
 '순운전자본비율',
 'CASHFLOW대총자본비율',
 '순운전자본회전률', 'label', '1주당순이익', '1주당정상영업이익']]

# KOSDAQ

train_df = pd.read_csv('./datasets/KOSDAQ_TRAIN_DATA.csv', index_col=0)
train_df = train_df.iloc[:,2:]
test_df = pd.read_csv('./datasets/KOSDAQ_TEST1_DATA.csv', index_col=0)
test_df = test_df.iloc[:,2:]

train_df = train_df[['총자본증가율',
 '총자본순이익률',
 '자기자본순이익률',
 '매출원가대매출액비율',
 '금융비용대총비용비율',
 '자기자본구성비율',
 '순운전자본비율',
 'CASHFLOW대총자본비율',
 '순운전자본회전률', 'label', '1주당순이익', '1주당정상영업이익']]

round(train_df[train_df['label'] == 1].describe().T,4)

round(train_df[train_df['label'] == 0].describe().T,4)
